Add files via upload
Added to readsnap function to identify whether the given snapshot is a single or multi-part file and generate an array of filenames to load. Currently,the file actually being loaded is simply the first file in the array, but this can be easily parallelized by passing an index based on the node id running the script.