Add files via upload
Changed to create a more general array of file names for both multiple snapshots and/or multipart files. You now must define a min and max snapshot in main, as well as a step (how many snapshots to skip). Should by default load multipart snapshots on the same node if the tasks per node is selected to be the same number of multi_part files.