Add files via upload
Parallel Shrinking Sphere algorithm to use with readsnap in order to find the center of galaxies. Compiles, but still needs to be debugged. Currently depends on knowing the number of files per snapshot, which will either need to be user defined or pulled from the headers. Currently will only work for snapshots with the same number of files per snapshot, but we can extend functionality later.